Include stable_sort_by in Array class
