+++ lib/more_core_extensions/all.rb
@@ -3,7 +3,6 @@ require 'more_core_extensions/version'
 require 'more_core_extensions/core_ext/array'
 require 'more_core_extensions/core_ext/benchmark'
 require 'more_core_extensions/core_ext/class'
-require 'more_core_extensions/core_ext/enumerable'
 require 'more_core_extensions/core_ext/hash'
 require 'more_core_extensions/core_ext/math'
 require 'more_core_extensions/core_ext/module'
