+++ lib/more_core_extensions/core_ext/array.rb
@@ -5,5 +5,6 @@ require 'more_core_extensions/core_ext/array/inclusions'
 require 'more_core_extensions/core_ext/array/math'
 require 'more_core_extensions/core_ext/array/nested'
 require 'more_core_extensions/core_ext/array/random'
+require 'more_core_extensions/core_ext/array/sorting'
 require 'more_core_extensions/core_ext/array/stretch'
 require 'more_core_extensions/core_ext/array/tableize'
