--- lib/more_core_extensions/core_ext/enumerable.rb
@@ -1,1 +1,0 @@
-require 'more_core_extensions/core_ext/enumerable/sorting'
